backend/requirements.txt
# --- Core Framework ---
Django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.2
django-cors-headers>=4.4
django-environ>=0.11

# --- Authentication & Permissions ---
djangorestframework-simplejwt>=5.3
google-auth>=2.29.0

# --- Admin & UI Enhancements ---
django-jazzmin>=3.0.0

# --- API Documentation (Swagger) ---
drf-yasg>=1.21.8
drf-spectacular>=0.27.0

# --- Media, Images, and File Handling ---
Pillow>=10.0.0

# --- Database (PostgreSQL ready for production) ---
psycopg2-binary>=2.9.9
dj-database-url>=2.1

# --- PDF and Reports ---
reportlab>=4.2.0

# --- WebSockets & Real-time (optional) ---
channels>=4.0.0
channels-redis>=4.1.0
redis>=5.0.0

# --- Developer Tools ---
django-extensions>=3.2.3

# --- Production (Render/Railway/Heroku) ---
whitenoise>=6.6.0
gunicorn>=21.2.0
